SPMG6T9-74-fix-accesskey-accessid
Hi guys, AWS was on our ass for placing access key and id in plain sight. I'm using b64 encrypted strings to replace this, then decrypt them again so they wont be in plain sight. Also, add a flask route to just a slash for easier testing; its annoying to land on 404 :D